requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
pillow
matplotlib
seaborn
plotly
pandas